pageturner/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
boringavatars==1.0.2
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==41.0.4
cssselect==1.2.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==4.2.5
django-allauth==0.60.1
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.3
django-formtools==2.4.1
django-import-export==3.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
environs==10.3.0
et-xmlfile==1.1.0
filelock==3.12.4
flake8==7.0.0
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
identify==2.5.36
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.20.2
mccabe==0.7.0
nodeenv==1.8.0
numpy==1.26.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
parsel==1.8.1
Pillow==10.0.1
platformdirs==4.2.1
plotly==5.17.0
pre-commit==3.7.1
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
PyDispatcher==2.0.7
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.35.1
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rpds-py==0.18.1
service-identity==23.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
tldextract==3.6.0
trio==0.22.2
trio-websocket==0.11.1
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
virtualenv==20.26.2
w3lib==2.1.2
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
